requirements.txt
python-telegram-bot==13.15
web3==6.0.0
python-dotenv==1.0.0
requests==2.28.2
python-dateutil==2.8.2
aiohttp==3.8.4
sqlite3-utils==0.1
pytz==2023.3